Crate plugins

Servo's compiler plugin/macro crate

Attributes this crate provides:

#[privatize] : Forces all fields in a struct/enum to be private

#[jstraceable] : Auto-derives an implementation of JSTraceable for a struct in the script crate

#[must_root] : Prevents data of the marked type from being used on the stack. See the lints module for more details

#[dom_struct] : Implies #[privatize],#[jstraceable], and #[must_root]. Use this for structs that correspond to a DOM type

Modules

casing

jstraceable
Handles the auto-deriving for #[jstraceable]

lints

reflector
Autogenerates implementations of Reflectable on DOM structs

utils
Utilities for writing plugins

Functions

plugin_registrar