Struct plugins::lints::unrooted_must_root::UnrootedPass

pub struct UnrootedPass;

Lint for ensuring safe usage of unrooted pointers

This lint (disable with -A unrooted-must-root/#[allow(unrooted_must_root)]) ensures that #[must_root] values are used correctly.

"Incorrect" usage includes:

Not being used in a struct/enum field which is not #[must_root] itself

Not being used as an argument to a function (Except onces named new and new_inherited)

Not being bound locally in a let statement, assignment, for loop, or match statement.

This helps catch most situations where pointers like JS<T> are used in a way that they can be invalidated by a GC pass.

Structs which have their own mechanism of rooting their unrooted contents (e.g. ScriptThread) can be marked as #[allow(unrooted_must_root)]. Smart pointers which root their interior type can be marked as #[allow_unrooted_interior]

Methods

impl UnrootedPass

fn new() -> UnrootedPass