Crate plugins

Servo's compiler plugin/macro crate

Attributes this crate provides:

#[privatize] : Forces all fields in a struct/enum to be private

#[derive(JSTraceable)] : Auto-derives an implementation of JSTraceable for a struct in the script crate

#[must_root] : Prevents data of the marked type from being used on the stack. See the lints module for more details

#[dom_struct] : Implies #[privatize],#[derive(JSTraceable)], and #[must_root]. Use this for structs that correspond to a DOM type

Modules

jstraceable
Handles the auto-deriving for #[derive(JSTraceable)]

lints

reflector
Autogenerates implementations of Reflectable on DOM structs

utils
Utilities for writing plugins

Functions

plugin_registrar

register_clippy